MOJO-1199 - Build number plugin to support GIT SCM
The build number plugin assumes that SVN is being used.
It would be nice to support GIT as this creates SHA1 revision numbers.
Attached is a patch to do this - the pom.xml will need to be adjusted to use the GIT SCM revision from maven when it is released.